+++ packages/vue-query/src/index.ts
@@ -1,9 +1,9 @@
 import { createRouterUtils } from './utils-router'
 
 export * from './key'
+export * from './procedure-utils'
 export * from './types'
 export * from './utils-general'
-export * from './utils-procedure'
 export * from './utils-router'
 
 export const createORPCVueQueryUtils = createRouterUtils
